Do not output the _many-to-one_ alignment except in `--m2m` mode.
The alignment files take significant amount of disk space.  Most users
of the standard mode are only interested in the _one-to-one_ alignments.
With that change, the intermediate alignment files are only kept in the
`--m2m` mode, which is expected to be selected only by users interested
in these alignments.

Closes #30.  Output is optional in the sense that one can always use
`--m2m` to get the _many-to-one_ alignment even if not interested in the
_one-to-many_.