Ignore errors probably caused by no alignments found
Closes #29

When aligning a target to a large number of queries, it is sometimes
impossible to know in advance if each of them is really alignable.
Ignoring the error is a much better outcome than interrupting the
pipeline and asking the user run the pipeline again after removing
the problematic genomes from the sample sheet (or running again
with a more sensitive seed).

Errors caused by lack of memory etc. are still errors.