Reduce verbosity of multiqc_assemblyscan_plot_data.
Pushing this commit now to trigger a new CI run of the nf-core branch
protection.

This said, multiqc_assemblyscan_plot_data combines one file per _query_
genome, and removing the `tag` ensure that the list of file names does
not clutter the screen when monitoring the pipeline run.  The nf-core
MultiQC module also does not have a `tag`.

Closes #64