Restore BED support, sort SAM/BAM/CRAM output.
Closes #56 for BED support.

SAM, BAM, and CRAM files are sorted for the convenience of the user,
who may for instance want to index them and use them as genome browser
tracks.

For the BAM and CRAM files, a sequence dictionary is computed before
export, to include all the sequences from the _target_ genome including
those that do not have an alignment to a given _query_, so that BAM or
CRAM files from multiple queries can be merged together without breaking
the sort order.

An assembly tag is also added, in case it is useful in the future.